modify: エラーページの名前変更、デザイン修正
